+++ deploy.sh
@@ -1,8 +1,30 @@
 #!/bin/bash
-npm ci
-npm install -g aws-cdk
+source ./local-vars.sh
 
-cd ./src/infrastructure || exit 1
+CDK_DEPLOY_ACCOUNT=$(aws sts get-caller-identity --profile "$AWS_PROFILE" | jq -r .Account)
+
+echo "Using project-local AWS CDK version..."
+npx cdk --version  # Prints the version being used
+
+if [ -z "$CDK_DEPLOY_ACCOUNT" ]; then
+  echo "AWS credentials not found. Did you run 'aws configure' or set AWS_PROFILE?"
+  #exit 1
+fi
+echo "AWS Account: $CDK_DEPLOY_ACCOUNT"
+echo "AWS Region: $AWS_REGION"
+
+cd ./src/infrastructure  || echo "Failed to change directory!" #; exit 1; }
+
+echo "Bootstrapping CDK Toolkit..."
+if ! aws cloudformation describe-stacks --stack-name CDKToolkit --region "$AWS_REGION" > /dev/null 2>&1; then
+  npx cdk bootstrap aws://"$CDK_DEPLOY_ACCOUNT"/"$AWS_REGION" || echo "Bootstrap failed!"; #exit 1
+else
+  echo  "CDK Toolkit stack already exists in $AWS_REGION. Skipping bootstrapping..."
+fi
+
+echo "Deploying all stacks..."
 npx cdk deploy --require-approval=never --ci --all
+echo "Deployment completed successfully."
+cd ../..
 
-exit 0
+#exit 0