+++ local-vars.sh
@@ -1,0 +1,7 @@
+export SERVICE=spod-knmn2m4
+export BUILD_STAGE=dev
+export AWS_REGION=us-east-1
+
+export AWS_PROFILE=SPod-profile
+
+#export DOMAIN=app.spod.com